Added glossaries, prepared for 0.1 release
This repository now shows the full example of how this will be used
within our other org repositories. The shared glossary will be
expanded over time, and now the common files can be centrally managed
and easily updated. xtask can include rustme if we desire, but I'm
currently envisioning this as a standalone installation.